Source/maizi-stu-003-project-dev/requirement.txt
Django==1.8.4
MySQL-python==1.2.3
Pillow==2.9.0
wheel==0.24.0
